# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListNotificationsResponse:
    properties:
      notifications:
        items:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.NotificationResponse'
        type: array
      total:
        type: integer
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.NotificationResponse:
    properties:
      body:
        type: string
      channel:
        type: string
      created_at:
        type: string
      id:
        type: string
      sent_at:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.SendPushRequest:
    properties:
      body:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      title:
        type: string
      user_id:
        type: string
    required:
    - body
    - title
    - user_id
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.SendSMSRequest:
    properties:
      body:
        type: string
      user_id:
        type: string
    required:
    - body
    - user_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for the gateway service of the School Tracking System.
  title: Gateway Service API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Retrieve a paginated list of notification logs
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by User ID
        in: query
        name: user_id
        type: string
      - description: Filter by Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListNotificationsResponse'
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}:
    get:
      description: Retrieve a notification log entry by UUID
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.NotificationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get notification details
      tags:
      - notifications
  /notifications/push:
    post:
      consumes:
      - application/json
      description: Enqueue a new push notification via FCM
      parameters:
      - description: Push Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.SendPushRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send a push notification
      tags:
      - notifications
  /notifications/retry:
    post:
      description: Manually trigger a retry for all notifications in 'failed' status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Retry failed notifications
      tags:
      - notifications
  /notifications/sms:
    post:
      consumes:
      - application/json
      description: Enqueue a new SMS via Twilio
      parameters:
      - description: SMS Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.SendSMSRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send an SMS notification
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"